major: Added support for environment variables. (#5)
This change affects how arguments and options are defined. The same information can be provided, but in a slightly different format (now in a struct, instead of as arguments to a function separately). This means that the number of configuration options can grow easier to allow the library to be more flexible, while maintaining backwards compatibility in the future, and while keeping configuration concise when working with simpler parameters.